cpp: qtwidgets: move shaders to versioned dir
